+++ apps/frontend/src/pages/routes/PublicRoutes.tsx
@@ -5,7 +5,11 @@ import { useUnit } from 'effector-react'
 import { userModel } from '@entities/user/model'
 
 export const PublicRoutes = () => {
-    const [user] = useUnit([userModel.stores.user])
+    const [user, sessionChecked] = useUnit([userModel.stores.user, userModel.stores.sessionChecked])
+
+    if (!sessionChecked) {
+        return null
+    }
 
     return !user ? <Outlet /> : <Navigate to="/" replace />
 }